refactor (channelDashboard) : channelDashboard ->  channel 폴더 명 수정
channelDashboard ->  channel 폴더 명 수정
channel 폴더 아래 파일 및 폴더 channel 단어 제거
