+++ app/(protected)/(channelLinked)/channel/page.ts
@@ -1,0 +1,1 @@
+export { ChannelPage as default } from '@/pages/channel'
