--- app/channelDashboard/[id]/page.ts
@@ -1,1 +1,0 @@
-export { ChannelDashboardPage as default } from '@/pages/channelDashboard'
