+++ src/entities/channel/kpiCard/index.ts
@@ -1,0 +1,3 @@
+export { KpiCard } from './ui/KpiCard'
+export type { KpiDto } from './model/types'
+export { mockKpi } from './mock/mockKpi'
