+++ src/entities/channel/kpiCard/mock/mockKpi.ts
@@ -1,0 +1,8 @@
+import type { KpiDto } from '../model/types'
+
+export const mockKpi: KpiDto = {
+  totalViews: 1000000,
+  avgEngagementRate: 42,
+  avgRetentionRate: 34,
+  weeklyUploadCount: 1.43,
+}
