+++ src/entities/channel/kpiCard/model/types.ts
@@ -1,0 +1,6 @@
+export interface KpiDto {
+  totalViews: number
+  avgEngagementRate: number
+  avgRetentionRate: number
+  weeklyUploadCount: number
+}
