+++ src/entities/channel/subscriberGrowth/model/types.ts
@@ -1,0 +1,8 @@
+export interface SubscriberGrowthPoint {
+  date: string
+  subscriberCount: number
+}
+
+export interface SubscriberGrowthDto {
+  points: SubscriberGrowthPoint[]
+}
