+++ src/entities/channel/trendingVideo/index.ts
@@ -1,0 +1,2 @@
+export { mockTrendingVideo } from './mock/mockTrendingVideo'
+export type { TrendingVideoItem } from './model/types'
