+++ src/entities/channel/trendingVideo/model/types.ts
