--- src/entities/channelDashboard/channelKpiCard/index.ts
@@ -1,3 +1,0 @@
-export { ChannelKpiCard } from './ui/ChannelKpiCard'
-export type { ChannelKpiDto } from './model/types'
-export { mockChannelKpi } from './mock/mockChannelKpi'
