--- src/entities/channelDashboard/channelKpiCard/mock/mockChannelKpi.ts
@@ -1,8 +1,0 @@
-import type { ChannelKpiDto } from '../model/types'
-
-export const mockChannelKpi: ChannelKpiDto = {
-  totalViews: 1000000,
-  avgEngagementRate: 42,
-  avgRetentionRate: 34,
-  weeklyUploadCount: 1.43,
-}
