--- src/entities/channelDashboard/channelKpiCard/model/types.ts
@@ -1,6 +1,0 @@
-export interface ChannelKpiDto {
-  totalViews: number
-  avgEngagementRate: number
-  avgRetentionRate: number
-  weeklyUploadCount: number
-}
