--- src/entities/channelDashboard/channelSubscriberGrowth/index.ts
@@ -1,6 +1,0 @@
-export type {
-  ChannelSubscriberGrowthDto,
-  SubscriberGrowthPoint,
-} from './model/types'
-export { mockChannelSubscriberGrowth } from './mock/mockChannelSubscriberGrowth'
-export { SubscriberGrowthChart } from './ui/SubscriberGrowthChart'
