--- src/entities/channelDashboard/channelSubscriberGrowth/model/types.ts
@@ -1,8 +1,0 @@
-export interface SubscriberGrowthPoint {
-  date: string
-  subscriberCount: number
-}
-
-export interface ChannelSubscriberGrowthDto {
-  points: SubscriberGrowthPoint[]
-}
