--- src/entities/channelDashboard/channelTrendingVideo/index.ts
@@ -1,2 +1,0 @@
-export { mockChannelTrendingVideo } from './mock/mockChannelTrendingVideo'
-export type { TrendingVideoItem } from './model/types'
