--- src/entities/channelDashboard/channelTrendingVideo/mock/mockChannelTrendingVideo.ts
@@ -1,55 +1,0 @@
-import type { TrendingVideoItem } from '../model/types'
-import mockThumbnailImage from '@/shared/assets/mock/mockThumbnailImage.png'
-
-export const mockChannelTrendingVideo: TrendingVideoItem[] = [
-  {
-    rank: 1,
-    title:
-      '제목입니다. 2026 최신 AI 폰 총정리 | 갤럭시 S26 vs 아이폰 17 비교2026 최신 AI 폰 총정리 | 갤럭시 S26 vs 아이폰 17 비교 2026 최신 AI 폰 총정리 | 갤럭시 S26 vs 아이폰 17 비교',
-    thumbnailUrl: mockThumbnailImage.src,
-    viewCount: 1000000,
-    engagementRate: 100,
-    ctr: 100,
-    retentionRate: 100,
-  },
-  {
-    rank: 2,
-    title:
-      '제목입니다. 2026 최신 AI 폰 총정리 | 갤럭시 S26 vs 아이폰 17 비교2026 최신 AI 폰 총정리 | 갤럭시 S26 vs 아이폰 17 비교 2026 최신 AI 폰 총정리 | 갤럭시 S26 vs 아이폰 17 비교',
-    thumbnailUrl: mockThumbnailImage.src,
-    viewCount: 1000000,
-    engagementRate: 100,
-    ctr: 100,
-    retentionRate: 100,
-  },
-  {
-    rank: 3,
-    title:
-      '제목입니다. 2026 최신 AI 폰 총정리 | 갤럭시 S26 vs 아이폰 17 비교2026 최신 AI 폰 총정리 | 갤럭시 S26 vs 아이폰 17 비교 2026 최신 AI 폰 총정리 | 갤럭시 S26 vs 아이폰 17 비교',
-    thumbnailUrl: mockThumbnailImage.src,
-    viewCount: 1000000,
-    engagementRate: 100,
-    ctr: 100,
-    retentionRate: 100,
-  },
-  {
-    rank: 4,
-    title:
-      '제목입니다. 2026 최신 AI 폰 총정리 | 갤럭시 S26 vs 아이폰 17 비교2026 최신 AI 폰 총정리 | 갤럭시 S26 vs 아이폰 17 비교 2026 최신 AI 폰 총정리 | 갤럭시 S26 vs 아이폰 17 비교',
-    thumbnailUrl: mockThumbnailImage.src,
-    viewCount: 1000000,
-    engagementRate: 100,
-    ctr: 100,
-    retentionRate: 100,
-  },
-  {
-    rank: 5,
-    title:
-      '제목입니다. 2026 최신 AI 폰 총정리 | 갤럭시 S26 vs 아이폰 17 비교2026 최신 AI 폰 총정리 | 갤럭시 S26 vs 아이폰 17 비교 2026 최신 AI 폰 총정리 | 갤럭시 S26 vs 아이폰 17 비교',
-    thumbnailUrl: mockThumbnailImage.src,
-    viewCount: 1000000,
-    engagementRate: 100,
-    ctr: 100,
-    retentionRate: 100,
-  },
-]
