+++ src/features/channel/kpi/index.ts
@@ -1,0 +1,1 @@
+export { useKpi } from './model/useKpi'
