+++ src/features/channel/subscriberGrowth/model/useSubscriberGrowth.ts
@@ -1,0 +1,10 @@
+import { useQuery } from '@tanstack/react-query'
+import { fetchSubscriberGrowth } from '../api/subscriberGrowthApi'
+
+export function useSubscriberGrowth(channelId: string, range: string) {
+  return useQuery({
+    queryKey: ['channel', channelId, range],
+    queryFn: () => fetchSubscriberGrowth(channelId, range),
+    enabled: !!channelId,
+  })
+}
