--- src/features/channelDashboard/channelKpi/index.ts
@@ -1,1 +1,0 @@
-export { useChannelKpi } from './model/useChannelKpi'
