+++ src/pages/channel/index.ts
@@ -1,0 +1,1 @@
+export { ChannelPage } from './ui/ChannelPage'
