--- src/pages/channelDashboard/index.ts
@@ -1,1 +1,0 @@
-export { ChannelDashboardPage } from './ui/ChannelDashboardPage'
