--- src/shared/api/msw/handlers/channelKpiHandlers.ts
@@ -1,16 +1,0 @@
-import { http, HttpResponse } from 'msw'
-
-import { mockChannelKpi } from '@/entities/channelDashboard/channelKpiCard'
-
-export const channelKpiHandlers = [
-  http.get(
-    `${process.env.NEXT_PUBLIC_API_URL}/channels/:channelId/kpi`,
-    async () => {
-      return HttpResponse.json({
-        success: true,
-        responseDto: mockChannelKpi,
-        error: null,
-      })
-    }
-  ),
-]
