+++ src/widgets/channel/Kpi/index.ts
@@ -1,0 +1,1 @@
+export { KpiSection } from './ui/KpiSection'
