--- src/widgets/channelDashboard/channelKpi/index.ts
@@ -1,1 +1,0 @@
-export { ChannelKpiSection } from './ui/ChannelKpiSection'
